+++ changelog.md
@@ -1,4 +1,8 @@
 # ACEmulator Change Log
+### 2017-10-06
+[Jyrus]
+* Moved WeenieType keys into their own class and implemented unlocking doors via keys
+
 ### 2017-10-04
 [Immortalbob]
 * Identified unknown enums 381 382 383 384 387 388 389 and updated acecharacter.cs and propertyint.cs.